Add new slideshow component
