+++ src/components/masthead/masthead.hbs
@@ -2,12 +2,12 @@
 By use of this code snippet, I agree to the Brightcove Publisher T and C
 found at https://accounts.brightcove.com/en/terms-and-conditions/.
 -->
-<article class="masthead masthead--{{type}} {{#if images}}masthead--has-images{{else}} masthead--no-images{{/if}}" data-lp-initial-images="{{masthead.json}}">
+<article class="masthead masthead--{{type}} {{#if images}}masthead--has-images{{else}} masthead--no-images{{/if}}">
   {{#if masthead.first}}
-    <div class="masthead__wrap">
-      <div class="masthead__images">
+    <div class="slideshow" data-lp-initial-images="{{masthead.json}}">
+      <div class="slideshow__images">
         {{#with masthead.first}}
-          <div class="masthead__slide masthead__slide--num_0" style="background-image: url({{large.hq}})"></div>
+          <div class="slideshow__slide slideshow__slide--num_0" style="background-image: url({{large.hq}})"></div>
         {{/with}}
       </div>
     </div>
