--- src/components/masthead/masthead_actions.js
@@ -1,12 +1,0 @@
-import Arkham from "../../core/arkham";
-
-let MastheadActions = {
-  advanceImage: () => {
-    Arkham.trigger("masthead.nextImage", {});
-  },
-  loopAround: () => {
-    Arkham.trigger("masthead.loopAround", {});
-  }
-};
-
-export default MastheadActions;
