+++ src/components/slideshow/index.js
@@ -1,0 +1,4 @@
+import Slideshow from "./slideshow";
+import "./slideshow.scss";
+
+export default Slideshow;
