+++ src/components/slideshow/slideshow.scss
@@ -1,0 +1,83 @@
+@import "../../../sass/webpack_deps";
+
+.slideshow {
+  height: 100%;
+
+  &__images {
+    position: relative;
+
+    height: 100%;
+  }
+
+  &__slide {
+    position: absolute;
+    z-index: 3;
+
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+
+    background: center no-repeat $h2-color;
+    background-size: cover;
+
+    transition: opacity $animation-speed-ui, transform $animation-speed-ui;
+  }
+
+  &--fade &__slide {
+    opacity: 1;
+
+    transition-duration: $animation-speed-ui*2;
+    transition-timing-function: ease-in-out;
+
+    &--num_-1 {
+      z-index: 1;
+      opacity: 1;
+    }
+
+    &--num_1 {
+      z-index: 5;
+      opacity: 0;
+    }
+  }
+
+  &--slide {
+    @include container();
+
+    @media (min-width: $min-1290) {
+      padding: 0;
+      max-width: 90%;
+    }
+  }
+
+  &--slide &__slide {
+    &--num_-2 {
+      transform: translateX(calc(-200% - #{$gutter * 2}));
+    }
+
+    &--num_-1 {
+      transform: translateX(calc(-100% - #{$gutter * 1}));
+    }
+
+    &--num_1 {
+      transform: translateX(calc(100% + #{$gutter * 1}));
+    }
+
+    &--num_2 {
+      transform: translateX(calc(200% + #{$gutter * 2}));
+    }
+
+    &--num_-2,
+    &--num_-1,
+    &--num_1,
+    &--num_2 {
+      opacity: .75;
+
+      cursor: pointer;
+
+      &:hover {
+        opacity: 1;
+      }
+    }
+  }
+}
